Can we pretend that airplanes in the night sky are like shooting stars ⭐
Typing in the closure is pretty mid compared to typescript, but those optimisations are cool as. Also not having an extra build step is nice. Also just fun having it lol.

Also downloads the closure compiler jar in the Makefile since I needed it on my windows PC and couldn't be assed installing it